Revamp IEnumerable implementations
